requirements.txt
transformers
torch
numpy
Pillow
streamlit
tensorflow